refactored tab width

+++ inc/graph.h
@@ -2,22 +2,21 @@
 
 #include <vector>
 #include <array>
-#include "imgui.h"
 
 struct SensorDataFrame{
-  float time;
-  std::array<float, 3> acceleration;
-  std::array<float, 3> rotation;
-  std::array<float, 3> magnetic;
+    double time;
+    std::array<double, 3> acceleration;
+    std::array<double, 3> rotation;
+    std::array<double, 3> magnetic;
 };
 
 class Graph {
- public:
-  void display();
-  void add_point(SensorDataFrame new_data);
- private:
-  std::array<std::vector<float>, 3> acceleration;
-  std::array<std::vector<float>, 3> rotation;
-  std::array<std::vector<float>, 3> magnetic;
+public:
+    void display();
+    void add_point(SensorDataFrame new_data);
+private:
+    std::array<std::vector<double>, 3> acceleration;
+    std::array<std::vector<double>, 3> rotation;
+    std::array<std::vector<double>, 3> magnetic;
 };
